Move Legacy-Item to its own file (#2111)
